transfert dans fill-random
